Organize dashboard files in folders
