+++ src/sections/dashboard/animeWidget/AnimeWidget.module.scss
